basic text rendering

+++ crates/anchor-kit-core/src/primitives/color.rs
@@ -1,9 +1,9 @@
 #[derive(Clone, Copy, Debug)]
 pub struct Color {
-    pub r: u32,
-    pub g: u32,
-    pub b: u32,
-    pub a: u32,
+    pub r: u8,
+    pub g: u8,
+    pub b: u8,
+    pub a: u8,
 }
 
 impl Color {
